refactor previews into smaller components rafctor to three frames
